Stricter numerical types (Integer, Long and Double) in schema inferring. Issue #162

+++ src/test/scala/com/crealytics/spark/v2/excel/ProjectionAndFilterPushdownSuite.scala
@@ -29,70 +29,70 @@ object ProjectionAndFilterPushdownSuite {
 
   /* No projection*/
   val expectedInferredSchema = StructType(List(
-    StructField("Day", DoubleType, true),
-    StructField("Month", DoubleType, true),
+    StructField("Day", IntegerType, true),
+    StructField("Month", IntegerType, true),
     StructField("Customer ID", StringType, true),
     StructField("Customer Name", StringType, true),
-    StructField("Standard Package", DoubleType, true),
-    StructField("Extra Option 1", DoubleType, true),
-    StructField("Extra Option 2", DoubleType, true),
-    StructField("Extra Option 3", DoubleType, true),
+    StructField("Standard Package", IntegerType, true),
+    StructField("Extra Option 1", IntegerType, true),
+    StructField("Extra Option 2", IntegerType, true),
+    StructField("Extra Option 3", IntegerType, true),
     StructField("Staff", StringType, true)
   ))
 
   val expectedDataInferSchema: util.List[Row] = List(
-    Row(1.0d, 12.0d, "CA869", "Phạm Uyển Trinh", null, null, 2200.0d, null, "Ella Fitzgerald"),
-    Row(1.0d, 12.0d, "CA870", "Nguyễn Liên Thảo", null, null, 2000.0d, 1350.0d, "Ella Fitzgerald"),
-    Row(1.0d, 12.0d, "CA871", "Lê Thị Nga", 17000.0d, null, null, null, "Ella Fitzgerald"),
-    Row(1.0d, 12.0d, "CA872", "Phan Tố Nga", null, null, 2000.0d, null, "Teresa Teng"),
-    Row(1.0d, 12.0d, "CA873", "Nguyễn Thị Teresa Teng", null, null, 1200.0d, null, "Jesse Thomas")
+    Row(1, 12, "CA869", "Phạm Uyển Trinh", null, null, 2200, null, "Ella Fitzgerald"),
+    Row(1, 12, "CA870", "Nguyễn Liên Thảo", null, null, 2000, 1350, "Ella Fitzgerald"),
+    Row(1, 12, "CA871", "Lê Thị Nga", 17000, null, null, null, "Ella Fitzgerald"),
+    Row(1, 12, "CA872", "Phan Tố Nga", null, null, 2000, null, "Teresa Teng"),
+    Row(1, 12, "CA873", "Nguyễn Thị Teresa Teng", null, null, 1200, null, "Jesse Thomas")
   ).asJava
 
   /* Subset of columns, same order*/
   val expectedProjectionInferredSchema_01 = StructType(List(
-    StructField("Day", DoubleType, true),
-    StructField("Month", DoubleType, true),
+    StructField("Day", IntegerType, true),
+    StructField("Month", IntegerType, true),
     StructField("Customer ID", StringType, true),
     StructField("Customer Name", StringType, true),
     StructField("Staff", StringType, true)
   ))
 
   val expectedProjectionDataInferSchema_01: util.List[Row] = List(
-    Row(1.0d, 12.0d, "CA869", "Phạm Uyển Trinh", "Ella Fitzgerald"),
-    Row(1.0d, 12.0d, "CA870", "Nguyễn Liên Thảo", "Ella Fitzgerald"),
-    Row(1.0d, 12.0d, "CA871", "Lê Thị Nga", "Ella Fitzgerald"),
-    Row(1.0d, 12.0d, "CA872", "Phan Tố Nga", "Teresa Teng"),
-    Row(1.0d, 12.0d, "CA873", "Nguyễn Thị Teresa Teng", "Jesse Thomas")
+    Row(1, 12, "CA869", "Phạm Uyển Trinh", "Ella Fitzgerald"),
+    Row(1, 12, "CA870", "Nguyễn Liên Thảo", "Ella Fitzgerald"),
+    Row(1, 12, "CA871", "Lê Thị Nga", "Ella Fitzgerald"),
+    Row(1, 12, "CA872", "Phan Tố Nga", "Teresa Teng"),
+    Row(1, 12, "CA873", "Nguyễn Thị Teresa Teng", "Jesse Thomas")
   ).asJava
 
   /* Subset of columns, out of order*/
   val expectedProjectionInferredSchema_02 = StructType(List(
     StructField("Staff", StringType, true),
-    StructField("Month", DoubleType, true),
-    StructField("Day", DoubleType, true),
+    StructField("Month", IntegerType, true),
+    StructField("Day", IntegerType, true),
     StructField("Customer ID", StringType, true),
     StructField("Customer Name", StringType, true),
-    StructField("Standard Package", DoubleType, true)
+    StructField("Standard Package", IntegerType, true)
   ))
 
   val expectedProjectionDataInferSchema_02: util.List[Row] = List(
-    Row("Ella Fitzgerald", 12.0d, 1.0d, "CA869", "Phạm Uyển Trinh", null),
-    Row("Ella Fitzgerald", 12.0d, 1.0d, "CA870", "Nguyễn Liên Thảo", null),
-    Row("Ella Fitzgerald", 12.0d, 1.0d, "CA871", "Lê Thị Nga", 17000.0d),
-    Row("Teresa Teng", 12.0d, 1.0d, "CA872", "Phan Tố Nga", null),
-    Row("Jesse Thomas", 12.0d, 1.0d, "CA873", "Nguyễn Thị Teresa Teng", null)
+    Row("Ella Fitzgerald", 12, 1, "CA869", "Phạm Uyển Trinh", null),
+    Row("Ella Fitzgerald", 12, 1, "CA870", "Nguyễn Liên Thảo", null),
+    Row("Ella Fitzgerald", 12, 1, "CA871", "Lê Thị Nga", 17000),
+    Row("Teresa Teng", 12, 1, "CA872", "Phan Tố Nga", null),
+    Row("Jesse Thomas", 12, 1, "CA873", "Nguyễn Thị Teresa Teng", null)
   ).asJava
 
   /* Filtering, with same schemas as in projection*/
   val expectedDataFilterInferSchema_01: util.List[Row] = List(
-    Row(4.0d, 12.0d, "CA883", "Phạm Thanh Mai", 7000.0d, 4000.0d, null, null, "Jesse Thomas"),
-    Row(4.0d, 12.0d, "CA884", "Hoàng Ngọc Hà", 8000.0d, null, null, null, "Teresa Teng"),
-    Row(4.0d, 12.0d, "CA885", "Lê Minh Ngọc", 15000.0d, 12000.0d, null, null, "Teresa Teng")
+    Row(4, 12, "CA883", "Phạm Thanh Mai", 7000, 4000, null, null, "Jesse Thomas"),
+    Row(4, 12, "CA884", "Hoàng Ngọc Hà", 8000, null, null, null, "Teresa Teng"),
+    Row(4, 12, "CA885", "Lê Minh Ngọc", 15000, 12000, null, null, "Teresa Teng")
   ).asJava
 
   val expectedDataFilterInferSchema_02: util.List[Row] = List(
-    Row(4.0d, 12.0d, "CA883", "Phạm Thanh Mai", 7000.0d, 4000.0d, null, null, "Jesse Thomas"),
-    Row(4.0d, 12.0d, "CA885", "Lê Minh Ngọc", 15000.0d, 12000.0d, null, null, "Teresa Teng")
+    Row(4, 12, "CA883", "Phạm Thanh Mai", 7000, 4000, null, null, "Jesse Thomas"),
+    Row(4, 12, "CA885", "Lê Minh Ngọc", 15000, 12000, null, null, "Teresa Teng")
   ).asJava
 }
